Split configurable sampler stuff into modules.
